stakevote WIP

+++ contracts/stakevote/stakevote.hpp
@@ -4,14 +4,14 @@
 
 #include "../../contract-shared-headers/dacdirectory_shared.hpp"
 #include "../../contract-shared-headers/eosdactokens_shared.hpp"
-#include "../../contract-shared-headers/safemath.hpp"
+#include "../../contract-shared-headers/safemath/safemath.hpp"
 
 using namespace eosio;
 using namespace eosdac;
 using namespace std;
 
-static constexpr auto time_divisor          = S{10}.ipow(8).to<int128_t>();
-static constexpr auto stake_duration_factor = S{10}.to<int128_t>();
+static constexpr auto time_divisor          = S{10}.ipow(8).to<double>();
+static constexpr auto stake_duration_factor = S{10}.to<double>();
 
 CONTRACT stakevote : public contract {
   public:
